pyproject.toml
[project]
name = "english-test-generator"
version = "0.1.0"
description = "Streamlit app for generating English knowledge tests"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.30.0",
    "fpdf2>=2.7.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "requests>=2.31.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
